+++ uninstall.sh
@@ -27,6 +27,7 @@ HELPERS=(
     "$PREFIX/bin/plasma-theme-master-helper-firefox"
     "$PREFIX/bin/plasma-theme-master-helper-konsole"
     "$PREFIX/bin/plasma-theme-master-helper-kitty"
+    "$PREFIX/bin/plasma-theme-master-helper-gtk4-reload"
     "$PREFIX/bin/plasma-theme-master-uninstall"
 )